updated per comments
