+++ src/js/providers/video-actions-mixin.js
@@ -1,5 +1,5 @@
-import { style } from 'utils/css';
-import { fitToBounds, fitVideoUsingTransforms } from 'utils/video-fit';
+import { style, transform } from 'utils/css';
+import { Browser, OS } from 'environment/environment'; 
 
 const VideoActionsMixin = {
     container: null,
@@ -23,6 +23,7 @@ const VideoActionsMixin = {
             return false;
         }
         let _videotag = this.video;
+        const fitVideoUsingTransforms = Browser.ie || (OS.iOS && OS.version.major < 9) || Browser.androidNative;
         let styles = {
             objectFit: null,
             width: null,
@@ -37,10 +38,25 @@ const VideoActionsMixin = {
             }
         }
         if (fitVideoUsingTransforms) {
-            styles = fitToBounds(_videotag, width, height, stretching, styles);
+            const x = -Math.floor(_videotag.videoWidth / 2 + 1);
+            const y = -Math.floor(_videotag.videoHeight / 2 + 1);
+            let scaleX = Math.ceil(width * 100 / _videotag.videoWidth) / 100;
+            let scaleY = Math.ceil(height * 100 / _videotag.videoHeight) / 100;
+            if (stretching === 'none') {
+                scaleX = scaleY = 1;
+            } else if (stretching === 'fill') {
+                scaleX = scaleY = Math.max(scaleX, scaleY);
+            } else if (stretching === 'uniform') {
+                scaleX = scaleY = Math.min(scaleX, scaleY);
+            }
+            styles.width = _videotag.videoWidth;
+            styles.height = _videotag.videoHeight;
+            styles.top = styles.left = '50%';
+            styles.margin = 0;
+            transform(_videotag,
+                'translate(' + x + 'px, ' + y + 'px) scale(' + scaleX.toFixed(2) + ', ' + scaleY.toFixed(2) + ')');
         } 
         style(_videotag, styles);
-        return false;
     },
 
     getContainer: function() {
@@ -65,3 +81,4 @@ const VideoActionsMixin = {
 };
 
 export default VideoActionsMixin;
+
