+++ src/js/utils/helpers.js
@@ -33,6 +33,7 @@ import {
     css,
     clearCss,
     style,
+    transform,
     getRgba
 } from 'utils/css';
 import { ajax, crossdomain } from 'utils/ajax';
@@ -71,6 +72,7 @@ const helpers = Object.assign({}, parser, validator, playerutils, {
     css,
     clearCss,
     style,
+    transform,
     getRgba,
     ajax,
     crossdomain,
